requirements.txt
Django>=5.2.7
djangorestframework>=3.14.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=41.0.0
dj-database-url>=2.1.0
